[core] Add ConfigManager type to core package and change ConfigManager from models to use it
