+++ src/models/configurationManager.go
@@ -1,33 +1,12 @@
 package models
 
 import (
-	"encoding/json"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"sync"
+	"github.com/fibercrypto/FiberCryptoWallet/src/core"
 
 	"github.com/therecipe/qt/qml"
 
 	qtcore "github.com/therecipe/qt/core"
 )
-
-const (
-	pathToConfigFromHome         = ".fiber/config.json"
-	pathToDefaultWalletsFromHome = ".skycoin/wallets"
-	localWallet                  = iota
-	remoteWallet
-)
-
-var (
-	configManager *ConfigManager
-	once          sync.Once
-)
-
-type walletSource struct {
-	sourceType int    `json:"type"`
-	source     string `json:"source"`
-}
 
 type WalletSource struct {
 	qtcore.QObject
@@ -35,91 +14,28 @@ type WalletSource struct {
 	_ string `property:"source"`
 }
 
-func getDefaultWalletSource() *walletSource {
-	ws := new(walletSource)
-	ws.sourceType = localWallet
-	walletsDir := filepath.Join(os.Getenv("HOME"), pathToDefaultWalletsFromHome)
-	ws.source = walletsDir
-	return ws
-
-}
-
 type ConfigManager struct {
 	qtcore.QObject
-	sourceList []*walletSource        `json:"SourceList"`
-	node       string                 `json:"Node"`
-	_          func()                 `constructor:"init"`
-	_          func() []*WalletSource `slot:"getSources"`
+	configManager *core.ConfigManager
+	_             func()                 `constructor:"init"`
+	_             func() []*WalletSource `slot:"getSources"`
 }
 
 func (cm *ConfigManager) init() {
-
-	cm = GetConfigManager()
+	qml.QQmlEngine_SetObjectOwnership(cm, qml.QQmlEngine__CppOwnership)
+	cm.configManager = core.GetConfigManager()
+	cm.ConnectGetSources(cm.getSources)
 
 }
-func GetConfigManager() *ConfigManager {
-	once.Do(func() {
-		cm := NewConfigManager(nil)
-		qml.QQmlEngine_SetObjectOwnership(cm, qml.QQmlEngine__CppOwnership)
-		cm.ConnectGetSources(cm.getSources)
-		if configFileExist() {
-			cm = loadConfigFromFile()
-		} else {
-			cm = getDefaultConfigManager()
-		}
-		configManager = cm
-	})
-	return configManager
-}
+
 func (cm *ConfigManager) getSources() []*WalletSource {
 	wltsSource := make([]*WalletSource, 0)
-	for _, wltS := range cm.sourceList {
+	for _, wltS := range cm.configManager.GetSources() {
 		newWltSource := NewWalletSource(nil)
 		qml.QQmlEngine_SetObjectOwnership(newWltSource, qml.QQmlEngine__CppOwnership)
-		newWltSource.SetSourceType(wltS.sourceType)
-		newWltSource.SetSource(wltS.source)
+		newWltSource.SetSourceType(wltS.GetType())
+		newWltSource.SetSource(wltS.GetSource())
 		wltsSource = append(wltsSource, newWltSource)
 	}
 	return wltsSource
-}
-
-func configFileExist() bool {
-	fileDir := getConfigFileDir()
-	if _, err := os.Stat(fileDir); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
-}
-
-func loadConfigFromFile() *ConfigManager {
-	cm := NewConfigManager(nil)
-	fileDir := getConfigFileDir()
-	dat, err := ioutil.ReadFile(fileDir)
-	if err != nil {
-		return getDefaultConfigManager()
-	}
-	err = json.Unmarshal(dat, cm)
-	if err != nil {
-		return getDefaultConfigManager()
-	}
-	return cm
-
-}
-
-func getDefaultConfigManager() *ConfigManager {
-	cm := NewConfigManager(nil)
-	cm.node = "http://stagin.node.skycoin.net"
-	cm.sourceList = []*walletSource{getDefaultWalletSource()}
-	jsonFormat, _ := json.Marshal(cm)
-	ioutil.WriteFile(getConfigFileDir(), jsonFormat, 0644)
-	return cm
-
-}
-
-func getConfigFileDir() string {
-	homeDir := os.Getenv("HOME")
-	fileDir := filepath.Join(homeDir, pathToConfigFromHome)
-	return fileDir
 }
